Move pipeline functions to file for both versions of test to import
